requirements.txt
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
rembg==2.0.59
pillow==11.0.0
requests==2.32.3
numpy==2.0.2
scikit-image==0.24.0
opencv-python-headless==4.10.0.84
